+++ packages/soda-heroui/CHANGELOG.md
@@ -1,5 +1,12 @@
 # soda-tanstack-form
 
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies
+    - deepsea-tools@5.23.1
+
 ## 0.1.0
 
 ### Minor Changes